Modifies Dockerfile
We were bundle installing in the repo and then copying those into the image as we built it, but that depends on gems built on the user's system and was failing for me before I realized I needed to bundle install with output set to vendor/cache. Instead, I've changed it so that the image itself bundle installs as it's being built via Dockerfile. This should be more consistent for others and especially if they don't know you can point bundle's output somewhere. It also means we won't need to configure CI/CD to be able to handle bundle installing with the right version of ruby and such.